requirements.txt
Babel==2.0
Django==1.6.1
MySQL-python==1.2.3
PAM==0.4.2
PyJWT==0.1.2
South==1.0
Twisted-Core==13.2.0
amqp==1.3.3
anyjson==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.13
celery==3.1.7
certifi==1.0.1
chardet==2.0.1
collectd==1.0
colorama==0.2.5
dj-database-url==0.2.2
django-appconf==0.6
django-celery==3.1.1
django-celery-with-redis==3.0
django-compressor==1.4
django-jsonify==0.2.1
django-mptt==0.6.0
django-polymorphic==0.5.6
django-redis==1.4.5
django-smtp-ssl==1.0
django-tagging==0.3.1
flower==0.8.3
futures==3.0.3
gevent==1.0.1
greenlet==0.4.7
gunicorn==18.0
hiredis==0.1.1
html5lib==0.999
httplib2==0.7.7
icalendar==3.6.1
irc==8.5.4
kombu==3.0.8
mock==1.0.1
psycogreen==1.0
pycurl==7.19.3
pyparsing==2.0.1
pyserial==2.6
#python-apt==0.9.3.5
python-dateutil==2.2
python-debian==0.1.21-nmu2ubuntu2
pytz==2015.4
redis==2.9.0
requests==2.7.0
simplejson==3.3.1
six==1.5.1
ssh-import-id==3.21
tornado==4.2.1
twilio==3.4.1
urllib3==1.7.1
wheel==0.24.0
whisper==0.9.12
wsgiref==0.1.2
zope.interface==4.0.5
python-gcm==0.2
